+++ R/forecasts.R
@@ -202,7 +202,15 @@ summary.sircovid_forecast <- function(object, ...,
   d
 }
 
-# Sum sampled model over compartments
+##' Take a sampled grid search produced by \code{\link{sample_grid_scan}} summarise trajectories 
+##' over compartmets
+##' 
+##' @title Summarise trajectories
+##'
+##' @param sample_grid_res Results from \code{\link{sample_grid_scan}}
+##' 
+##' @export
+##' 
 sum_over_compartments <- function(sample_grid_res) {
   index <- odin_index(sample_grid_res$inputs$model$odin_model(
       user = sample_grid_res$inputs$model_params))
@@ -211,7 +219,7 @@ sum_over_compartments <- function(sample_grid_res) {
   n <- vapply(index, min, numeric(1)) - 1
   keep <- names(which(n > 0 & n <= ncol(sample_grid_res$trajectories)))
   f <- function(k) {
-    apply(sample_grid_res$trajectories[, k - 1, ], c(1, 3), sum)
+    apply(sample_grid_res$trajectories[, k - 1, , drop = FALSE], c(1, 3), sum)
   }
   totals <- lapply(index[keep], f)